fix: move backfill deprecated rule to 6_8
