+++ tests/migration/Core/V6_7/Migration1773392534MigrateDigitalDownloadRuleToProductTypeTest.php
@@ -8,7 +8,7 @@ use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelLifecycleManager;
 use Shopware\Core\Framework\Uuid\Uuid;
-use Shopware\Core\Migration\V6_7\Migration1773392534MigrateDigitalDownloadRuleToProductType;
+use Shopware\Core\Migration\V6_8\Migration1773392534MigrateDigitalDownloadRuleToProductType;
 
 /**
  * @internal
